change RequireOne name to RequireExactlyOne
