+++ test-d/require-exactly-one.ts
@@ -1,5 +1,5 @@
 import {expectType, expectError} from 'tsd';
-import {RequireOne} from '..';
+import {RequireExactlyOne} from '..';
 
 type SystemMessages = {
 	default: string;
@@ -10,7 +10,7 @@ type SystemMessages = {
 	optional?: string;
 };
 
-type ValidMessages = RequireOne<SystemMessages, 'macos' | 'linux'>;
+type ValidMessages = RequireExactlyOne<SystemMessages, 'macos' | 'linux'>;
 const test = (_: ValidMessages): void => {};
 
 test({macos: 'hey', default: 'hello'});
@@ -19,6 +19,6 @@ test({linux: 'sup', optional: 'howdy', default: 'hello'});
 expectError(test({}));
 expectError(test({macos: 'hey', linux: 'sup', default: 'hello'}));
 
-declare const oneWithoutKeys: RequireOne<{a: number; b: number}>;
+declare const oneWithoutKeys: RequireExactlyOne<{a: number; b: number}>;
 expectType<{a: number} | {b: number}>(oneWithoutKeys);
 expectError(expectType<{a: number; b: number}>(oneWithoutKeys));
